requirements.txt
aniso8601==9.0.1
attrs==23.1.0
blinker==1.6.2
cffi==1.15.1
click==8.1.7
Flask==2.0.2
flask-marshmallow==0.1.0
SQLAlchemy==1.4.49
flask-restx==0.5.1
pip~=23.2.1
wheel~=0.41.0
Jinja2~=3.1.2
MarkupSafe~=2.1.3
werkzeug~=2.3.7
python-dotenv~=1.0.0
itsdangerous~=2.1.2
greenlet~=2.0.2
setuptools~=68.0.0
packaging~=23.1
marshmallow~=3.20.1
Whoosh==2.6.0
autopep8==1.5.7
decorator==5.0.9
flasgger==0.9.5
Flask-JWT-Extended==4.2.3
Flask-SQLAlchemy==2.5.1
gunicorn==20.1.0
importlib-metadata==4.6.1
jsonschema==3.2.0
mistune==0.8.4
pycodestyle==2.7.0
PyJWT==2.1.0
PyMySQL==1.0.2
pyrsistent==0.18.0
six==1.16.0
toml==0.10.2
typing-extensions==3.10.0.0
validators==0.18.2
zipp==3.5.0